Create `ComputedFolder`` base class for `MultibranchWorkflowJob` and `OrganizationFolderJob`
- Moved `triggers` DSL method to `ComputedFolder`
- Moved `orphanedItemStrategy` DSL method to `ComputedFolder`
