+++ job-dsl-core/src/main/groovy/javaposse/jobdsl/dsl/jobs/OrganizationFolderJob.groovy
@@ -1,17 +1,15 @@
 package javaposse.jobdsl.dsl.jobs
 
+import javaposse.jobdsl.dsl.ComputedFolder
 import javaposse.jobdsl.dsl.ContextHelper
 import javaposse.jobdsl.dsl.DslContext
-import javaposse.jobdsl.dsl.Folder
 import javaposse.jobdsl.dsl.JobManagement
-import javaposse.jobdsl.dsl.helpers.triggers.MultibranchWorkflowTriggerContext
-import javaposse.jobdsl.dsl.helpers.workflow.OrphanedItemStrategyContext
 import javaposse.jobdsl.dsl.helpers.workflow.ScmNavigatorsContext
 
 /**
  * @since 1.56
  */
-class OrganizationFolderJob extends Folder {
+class OrganizationFolderJob extends ComputedFolder {
 
     protected OrganizationFolderJob(JobManagement jobManagement, String name) {
         super(jobManagement, name)
@@ -34,42 +32,6 @@ class OrganizationFolderJob extends Folder {
             Node navigators = project / navigators
             context.scmNavigatorNodes.each {
                 navigators << it
-            }
-        }
-    }
-
-    /**
-     * Sets the orphaned item strategy.
-     * @since 1.56
-     */
-    void orphanedItemStrategy(@DslContext(OrphanedItemStrategyContext) Closure closure) {
-        OrphanedItemStrategyContext context = new OrphanedItemStrategyContext()
-        ContextHelper.executeInContext(closure, context)
-
-        if (context.orphanedItemStrategyNode != null) {
-            configure { Node project ->
-                Node orphanedItemStrategy = project / 'orphanedItemStrategy'
-                if (orphanedItemStrategy) {
-                    // there can only be only one orphanedItemStrategy, so remove if there
-                    project.remove(orphanedItemStrategy)
-                }
-
-                project << context.orphanedItemStrategyNode
-            }
-        }
-    }
-
-    /**
-     * Sets the build triggers for this job.
-     * @since 1.56
-     */
-    void triggers(@DslContext(MultibranchWorkflowTriggerContext) Closure closure) {
-        MultibranchWorkflowTriggerContext context = new MultibranchWorkflowTriggerContext(jobManagement, this)
-        ContextHelper.executeInContext(closure, context)
-
-        configure { Node project ->
-            context.triggerNodes.each {
-                project / 'triggers' << it
             }
         }
     }
